refactor: extract get-cell-id to own file
