+++ src/shared/stores/get-cell-id.js
@@ -1,0 +1,39 @@
+import {
+    getContextSelectionId,
+    parseContextSelectionId,
+} from '../use-context-selection/use-context-selection.js'
+
+export const getCellId = ({
+    contextSelectionId,
+    item: { categoryOptionCombo, dataElement },
+}) => `${contextSelectionId}_${dataElement}_${categoryOptionCombo}`
+
+export const parseCellId = (cellId) => {
+    const [contextSelectionId, dataElement, categoryOptionCombo] =
+        cellId.split('_')
+    const parsedContextSelectionId = parseContextSelectionId(contextSelectionId)
+    return {
+        ...parsedContextSelectionId,
+        dataElement,
+        categoryOptionCombo,
+    }
+}
+
+export const getCellIdFromDataValueParams = (params) => {
+    if (!params) {
+        return null
+    }
+    const contextSelectionId = getContextSelectionId({
+        attributeOptions: params.co?.split(';'),
+        dataSetId: params.ds,
+        orgUnitId: params.ou,
+        periodId: params.ps,
+    })
+    return getCellId({
+        contextSelectionId,
+        item: { categoryOptionCombo: params.co, dataElement: params.de },
+    })
+}
+
+export const getCellIdFromMutationKey = (mutationKey) =>
+    getCellIdFromDataValueParams(mutationKey[1])
