+++ src/shared/stores/index.js
@@ -3,3 +3,5 @@ export { useValidationStore } from './validation-store.js'
 export { useHighlightedFieldStore } from './highlighted-field-store.js'
 export { useEntryFormStore } from './entry-form-store.js'
 export { useUnsavedDataStore } from './unsaved-data-store.js'
+export { useSyncErrorsStore } from './sync-errors-store.js'
+export * from './get-cell-id.js'
