+++ src/shared/stores/sync-errors-store.js
@@ -1,8 +1,5 @@
 import create from 'zustand'
-import {
-    getContextSelectionId,
-    parseContextSelectionId,
-} from '../use-context-selection/use-context-selection.js'
+import { parseCellId, getCellIdFromMutationKey } from './get-cell-id.js'
 
 const inititalState = {
     // keyed by cellId, see getCellId
@@ -18,13 +15,13 @@ export const useSyncErrorsStore = create((set, get) => ({
             return { comments }
         }),
     setErrorByMutationKey: (mutationKey, value) => {
-        const cellId = getCellIdFromDataValueParams(mutationKey[1])
+        const cellId = getCellIdFromMutationKey(mutationKey)
         get().setError(cellId, value)
     },
     getErrors: () => get().errors,
     getError: (cellId) => get().errors[cellId],
     getErrorByMutationKey: (mutationKey) =>
-        get().getError(getCellIdFromDataValueParams(mutationKey[1])),
+        get().getError(getCellIdFromMutationKey(mutationKey)),
     getErrorsForSelection: (contextSelectionId) => {
         const errors = get().getErrors()
         return Object.entries(errors)
@@ -38,32 +35,3 @@ export const useSyncErrorsStore = create((set, get) => ({
             })
     },
 }))
-
-export const getCellId = ({
-    contextSelectionId,
-    item: { categoryOptionCombo, dataElement },
-}) => `${contextSelectionId}_${dataElement}_${categoryOptionCombo}`
-
-const parseCellId = (cellId) => {
-    const [contextSelectionId, dataElement, categoryOptionCombo] =
-        cellId.split('_')
-    const parsedContextSelectionId = parseContextSelectionId(contextSelectionId)
-    return {
-        ...parsedContextSelectionId,
-        dataElement,
-        categoryOptionCombo,
-    }
-}
-
-export const getCellIdFromDataValueParams = (params) => {
-    const contextSelectionId = getContextSelectionId({
-        attributeOptions: params.co.split(';'),
-        dataSetId: params.ds,
-        orgUnitId: params.ou,
-        periodId: params.ps,
-    })
-    return getCellId({
-        contextSelectionId,
-        item: { categoryOptionCombo: params.co, dataElement: params.de },
-    })
-}
